+++ src/components/burger-constructor/burger-constructor-ingredient/burger-constructor-ingredient.module.css
@@ -1,0 +1,4 @@
+.ingredient{
+    display: flex; 
+    align-items: center;
+}
