+++ src/components/burger-constructor/burger-constructor.jsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState, useRef, useCallback } from 'react';
-import { ConstructorElement, Button, CurrencyIcon, DragIcon } from '@ya.praktikum/react-developer-burger-ui-components';
+import React, { useEffect, useState} from 'react';
+import { ConstructorElement, Button, CurrencyIcon} from '@ya.praktikum/react-developer-burger-ui-components';
 import styles from './burger-constructor.module.css';
 import PropTypes from 'prop-types';
 import { dataPropTypes } from '../utils/data-prop-types.js';
@@ -7,10 +7,11 @@ import Modal from '../modal/modal.jsx';
 import OrderDetails from '../order-details/order-details.jsx';
 import { useDispatch, useSelector } from 'react-redux';
 import { getIngredientsFromConstructor } from '../../services/selectors.js';
-import { useDrop, useDrag } from "react-dnd";
-import { ADD_INGREDIENT, SET_BUN, SET_TOTAL, DELETE_INGREDIENT, SWAP_INGREDIENTS} from '../../services/actions/burger-constructor.js'
+import { useDrop} from "react-dnd";
+import { SET_BUN, SET_TOTAL, DELETE_INGREDIENT} from '../../services/actions/burger-constructor.js'
 import {createOrderAction} from '../../services/actions/create-order.js';
-
+import {createUniqKeyForIngredientAction} from '../../services/actions/burger-constructor.js'
+import BurgerConstructorIngredient from './burger-constructor-ingredient/burger-constructor-ingredient.jsx';
 
 function BurgerConstructor() {
     const { ingredients, bun, total } = useSelector(getIngredientsFromConstructor);
@@ -37,7 +38,7 @@ function BurgerConstructor() {
     const [, dropTargetIngredient] = useDrop({
         accept: ["sauce", "main"],
         drop(item) {
-            dispatch({ type: ADD_INGREDIENT, item: item });
+           dispatch(createUniqKeyForIngredientAction(item));
         },
     });
 
@@ -72,22 +73,6 @@ function BurgerConstructor() {
     };
     
 
-      const ref = useRef(null);
-     
-      const [, drag] = useDrag({
-          type: "sort",
-          item: ingredients.index
-      });
-  
-      const [, drop] = useDrop({
-          accept: "sort",
-          drop(item) {
-              if (index !== ingredients) {
-                  dispatch({ type: SWAP_INGREDIENTS, index1: index, index2: item.index });
-              }
-          }
-      });
-      drag(drop(ref));
     return (
         <section className={styles.section}>
             <div className='mt-25 ml-8'>
@@ -107,19 +92,9 @@ function BurgerConstructor() {
 
                 <ul className={styles.scroll} ref={dropTargetIngredient} >
                     {ingredients && ingredients.length > 0 ? ingredients.map((item, index) => (
-                        <li key={index} className={`${styles.ingredient} mt-4`}   
-                        ref={ref}>
-                            <DragIcon type="primary" />
-                            <ConstructorElement
-                                text={item.name}
-                                price={item.price}
-                                thumbnail={item.image}
-                                handleClose={() => deleteIngredient(index)}
-                            //    handleDrop={handleDrop}
-                              
-                                                       
-                            />
-                        </li>)) :
+                    <BurgerConstructorIngredient key={item.key} item={item} index={index} onDelete={deleteIngredient}  />   
+                    
+                    )) :
                         (<div className={`${styles["empty-element"]} constructor-element constructor-element ml-8`}>
                             <div className={`${styles["empty-element-text"]} text text_type_main-default`}>Перетащите ингредиенты</div>
                         </div>)
