+++ src/services/actions/burger-constructor.js
@@ -1,5 +1,16 @@
 export const SET_BUN = "SET_BUN";
 export const ADD_INGREDIENT = "ADD_INGREDIENT";
-export const DELETE_INGREDIENT = "DELETE_INGREDIENT";
+export const DELETE_INGREDIENT = "DELETE_INGREDIENTS";
 export const SET_TOTAL = "SET_TOTAL";
-export const SWAP_INGREDIENTS = "SWAP_INGREDIENT";
+export const SWAP_INGREDIENT = "SWAP_INGREDIENT";
+
+
+export function createUniqKeyForIngredientAction(ingredients){
+    return function(dispatch) {
+        dispatch({
+            type: ADD_INGREDIENT,
+            item: {...ingredients,  key:Math.random()}
+           
+        });
+    }
+}
